pyproject.toml
[tool.poetry]
name = "psa-car-controller"
version = "3.0.0"
description = "This is a python program to control a psa car with connected_car v4 api."
authors = ["Florian Bezannier <florian.bezannier@hotmail.fr>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.7.0"
paho-mqtt = ">=1.5.0"
dash = ">=2"
dash-daq = "^0.5.0"
plotly = ">=5"
cryptography = ">=2.6"
Werkzeug = ">=1.0.0"
Flask = ">=1.0.4"
dash-bootstrap-components = ">=1"
ConfigUpdater = ">=3.0"
pandas = ">=0.23"
oauth2-client = "^1.2.1"
requests = "^2.27.1"
pytz = "^2021.3"
argparse = "^1.4.0"
geojson = "^2.5.0"
reverse-geocode = "^1.4.1"
androguard = "^3.3.5"
pycryptodomex = "^3.14.1"
pydantic = "^1.9.0"
"ruamel.yaml" = ">=0.15.0"
certifi = ">=14.05.14"
six = ">=1.10"
python-dateutil = ">=2.5.3"
urllib3 = ">=1.15.1"

[tool.poetry.dev-dependencies]
prospector = ">=1.3.0"
pre-commit = "^2.17.0"
coverage = "^6.3.2"
deepdiff = "^5.7.0"
selenium = "^4.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.autopep8]
max_line_length = 120
in-place = true
recursive = true
aggressive = 3
exclude = "psa_car_controller/__main__.py,psa_car_controller/psa/connected_car_api"